+++ tests/test_admin.rs
@@ -5,14 +5,14 @@ use std::time::Duration;
 use backoff::{ExponentialBackoff, Operation};
 
 use rdkafka::admin::{
-    AdminClient, AdminOptions, AlterConfig, ConfigEntry, ConfigSource, NewPartitions, NewTopic,
-    OwnedResourceSpecifier, ResourceSpecifier, TopicReplication,
+    AdminClient, AdminOptions, AlterConfig, ConfigEntry, ConfigSource, GroupResult, NewPartitions,
+    NewTopic, OwnedResourceSpecifier, ResourceSpecifier, TopicReplication,
 };
 use rdkafka::client::DefaultClientContext;
-use rdkafka::consumer::{BaseConsumer, Consumer, DefaultConsumerContext};
+use rdkafka::consumer::{BaseConsumer, CommitMode, Consumer, DefaultConsumerContext};
 use rdkafka::error::{KafkaError, RDKafkaErrorCode};
 use rdkafka::metadata::Metadata;
-use rdkafka::ClientConfig;
+use rdkafka::{ClientConfig, TopicPartitionList};
 
 use crate::utils::*;
 
@@ -28,6 +28,45 @@ fn create_admin_client() -> AdminClient<DefaultClientContext> {
     create_config()
         .create()
         .expect("admin client creation failed")
+}
+
+async fn create_consumer_group(consumer_group_name: &str) {
+    let admin_client = create_admin_client();
+    let topic_name = &rand_test_topic();
+    let consumer: BaseConsumer = create_config()
+        .set("group.id", consumer_group_name.clone())
+        .create()
+        .expect("create consumer failed");
+
+    admin_client
+        .create_topics(
+            &[NewTopic {
+                name: topic_name,
+                num_partitions: 1,
+                replication: TopicReplication::Fixed(1),
+                config: vec![],
+            }],
+            &AdminOptions::default(),
+        )
+        .await
+        .expect("topic creation failed");
+    let topic_partition_list = {
+        let mut lst = TopicPartitionList::new();
+        lst.add_partition(topic_name, 0);
+        lst
+    };
+    consumer
+        .assign(&topic_partition_list)
+        .expect("assign topic partition list failed");
+    consumer
+        .fetch_metadata(None, Duration::from_secs(3))
+        .expect("unable to fetch metadata");
+    consumer
+        .store_offset(topic_name, 0, -1)
+        .expect("store offset failed");
+    consumer
+        .commit_consumer_state(CommitMode::Sync)
+        .expect("commit the consumer state failed");
 }
 
 fn fetch_metadata(topic: &str) -> Metadata {
@@ -378,6 +417,56 @@ async fn test_configs() {
         .await
         .expect("alter configs failed");
     assert_eq!(res, &[Ok(OwnedResourceSpecifier::Broker(0))]);
+}
+
+#[tokio::test]
+async fn test_groups() {
+    let admin_client = create_admin_client();
+
+    // Verify that a valid group can be deleted.
+    {
+        let group_name = rand_test_group();
+        create_consumer_group(&group_name).await;
+        let res = admin_client
+            .delete_groups(&[&group_name], &AdminOptions::default())
+            .await;
+        assert_eq!(res, Ok(vec![Ok(group_name.to_string())]));
+    }
+
+    // Verify that attempting to delete an unknown group returns a "group not
+    // found" error.
+    {
+        let unknown_group_name = rand_test_group();
+        let res = admin_client
+            .delete_groups(&[&unknown_group_name], &AdminOptions::default())
+            .await;
+        let expected: GroupResult = Err((unknown_group_name, RDKafkaErrorCode::GroupIdNotFound));
+        assert_eq!(res, Ok(vec![expected]));
+    }
+
+    // Verify that deleting a valid and invalid group results in a mixed result
+    // set.
+    {
+        let group_name = rand_test_group();
+        let unknown_group_name = rand_test_group();
+        create_consumer_group(&group_name).await;
+        let res = admin_client
+            .delete_groups(
+                &[&group_name, &unknown_group_name],
+                &AdminOptions::default(),
+            )
+            .await;
+        assert_eq!(
+            res,
+            Ok(vec![
+                Ok(group_name.to_string()),
+                Err((
+                    unknown_group_name.to_string(),
+                    RDKafkaErrorCode::GroupIdNotFound
+                ))
+            ])
+        );
+    }
 }
 
 // Tests whether each admin operation properly reports an error if the entire