+++ rdkafka-sys/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+* Add the `RDKafkaDeleteGroup` and `RDKafkaGroupResult` type aliases to the
+  `types` module.
+
 ## v4.3.0+1.9.2 (2022-10-29)
 
 * Upgrade to librdkafka v1.9.2.